+++ packages/sdk-statement/src/api.ts
@@ -1,26 +1,60 @@
 import { HexString } from "@polkadot-api/substrate-bindings"
-import { fromHex } from "@polkadot-api/utils"
-import { SubmitResult } from "./types"
+import {
+  SubmitResult,
+  TopicFilter,
+  SubscriptionCallback,
+  Unsubscribe,
+  extractStatementEvent,
+} from "./types"
 
 export type RequestFn = <Reply = any, Params extends Array<any> = any[]>(
   method: string,
   params: Params,
 ) => Promise<Reply>
 
+export type FollowSubscription = (
+  subscriptionId: string,
+  handlers: { next: (event: unknown) => void; error: (error: Error) => void },
+) => void
+
+export type SubscriptionRequestFn = <SubscriptionId extends string>(
+  method: string,
+  params: unknown[],
+  handlers: {
+    onSuccess: (
+      subscriptionId: SubscriptionId,
+      followSubscription: FollowSubscription,
+    ) => void
+    onError: (error: Error) => void
+  },
+) => Unsubscribe
+
 export const getApi = (req: RequestFn) => ({
   submit: (stmt: HexString) =>
     req<SubmitResult | undefined, [HexString]>("statement_submit", [stmt]),
-
   dump: () => req<HexString[], []>("statement_dump", []),
+})
 
-  broadcasts: (matchAllTopics: HexString[]) =>
-    req<HexString[], [number[][]]>("statement_broadcastsStatement", [
-      matchAllTopics.map((v) => [...fromHex(v)]),
-    ]),
-
-  posted: (matchAllTopics: HexString[], dest: HexString) =>
-    req<HexString[], [number[][], number[]]>("statement_postedStatement", [
-      matchAllTopics.map((v) => [...fromHex(v)]),
-      [...fromHex(dest)],
-    ]),
+export const getSubscriptionApi = (req: SubscriptionRequestFn) => ({
+  subscribe: (
+    filter: TopicFilter,
+    onEvent: SubscriptionCallback,
+    onError?: (error: Error) => void,
+  ): Unsubscribe =>
+    req<string>("statement_subscribeStatement", [filter], {
+      onSuccess: (_, followSubscription) => {
+        followSubscription(_, {
+          next: (event: unknown) => {
+            if (typeof event === "string") {
+              onEvent({ statements: [event] })
+              return
+            }
+            const statementEvent = extractStatementEvent(event)
+            if (statementEvent) onEvent(statementEvent)
+          },
+          error: (e) => onError?.(e),
+        })
+      },
+      onError: (e) => onError?.(e),
+    }),
 })